move serverqueue related code out of branch.java
